fix : zipbap/api -> zipbap/app으로 폴더명 수정
