+++ HELP.md
@@ -1,7 +1,7 @@
 # Read Me First
 The following was discovered as part of building this project:
 
-* The original package name 'zipbap.zipbap-api' is invalid and this project uses 'zipbap.api' instead.
+* The original package name 'zipbap.zipbap-api' is invalid and this project uses 'zipbap.app' instead.
 
 # Getting Started
