--- src/main/kotlin/zipbap/api/ZipbapApiApplication.kt
@@ -1,13 +1,0 @@
-package zipbap.api
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@SpringBootApplication
-@EnableJpaAuditing
-class ZipbapApiApplication
-
-fun main(args: Array<String>) {
-	runApplication<ZipbapApiApplication>(*args)
-}
