--- src/main/kotlin/zipbap/api/domain/category/cookingtype/CookingType.kt
@@ -1,18 +1,0 @@
-package zipbap.api.domain.category.cookingtype
-
-import jakarta.persistence.*
-import zipbap.api.domain.base.BaseEntity
-
-@Entity
-@Table(name = "category_cooking_type")
-class CookingType(
-
-        @Column(name = "type", nullable = false,
-                length = 50, unique = true)
-        val type: String,
-
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        val id: Long? = null
-) : BaseEntity() {
-}
