--- src/main/kotlin/zipbap/api/domain/category/headcount/Headcount.kt
@@ -1,17 +1,0 @@
-package zipbap.api.domain.category.headcount
-
-import jakarta.persistence.*
-import zipbap.api.domain.base.BaseEntity
-
-@Entity
-@Table(name = "category_headcount")
-class Headcount(
-
-        @Column(name = "headcount", nullable = false,
-                length = 50, unique = true)
-        var headcount: String,
-
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        val id: Long? = null
-) : BaseEntity()
