--- src/main/kotlin/zipbap/api/domain/category/method/Method.kt
@@ -1,17 +1,0 @@
-package zipbap.api.domain.category.method
-
-import jakarta.persistence.*
-import zipbap.api.domain.base.BaseEntity
-
-@Entity
-@Table(name = "category_method")
-class Method(
-
-        @Column(name = "method", nullable = false,
-                length = 50, unique = true)
-        var method: String,
-
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        val id: Long? = null
-) : BaseEntity()
