--- src/main/kotlin/zipbap/api/domain/recipe/RecipeStatus.kt
@@ -1,5 +1,0 @@
-package zipbap.api.domain.recipe
-
-enum class RecipeStatus {
-    TEMPORARY, ACTIVE
-}
