--- src/main/kotlin/zipbap/api/domain/user/SocialType.kt
@@ -1,5 +1,0 @@
-package zipbap.api.domain.user
-
-enum class SocialType {
-    KAKAO, APPLE
-}
