--- src/main/kotlin/zipbap/api/global/code/BaseCode.kt
@@ -1,6 +1,0 @@
-package zipbap.api.global.code
-
-interface BaseCode {
-    val reasonDto: ReasonDto
-    val reasonHttpStatus: ReasonDto
-}
