--- src/main/kotlin/zipbap/api/global/code/BaseErrorCode.kt
@@ -1,6 +1,0 @@
-package zipbap.api.global.code
-
-interface BaseErrorCode {
-    val reason: ErrorReasonDto
-    val reasonHttpStatus: ErrorReasonDto
-}
