--- src/main/kotlin/zipbap/api/global/code/ErrorReasonDto.kt
@@ -1,11 +1,0 @@
-package zipbap.api.global.code
-
-import org.springframework.http.HttpStatus
-
-data class ErrorReasonDto(
-    val httpStatus: HttpStatus,
-    val isSuccess: Boolean,
-    val code: String,
-    val message: String
-) {
-}
