--- src/main/kotlin/zipbap/api/global/code/ReasonDto.kt
@@ -1,11 +1,0 @@
-package zipbap.api.global.code
-
-import org.springframework.http.HttpStatus
-
-data class ReasonDto(
-     val httpStatus: HttpStatus? = null,
-     val isSuccess: Boolean = false,
-     val code: String? = null,
-     val message: String? = null
-) {
-}
