+++ src/main/kotlin/zipbap/app/ZipbapApiApplication.kt
@@ -1,0 +1,11 @@
+package zipbap.app
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class ZipbapApiApplication
+
+fun main(args: Array<String>) {
+	runApplication<ZipbapApiApplication>(*args)
+}
