+++ src/main/kotlin/zipbap/app/global/code/BaseCode.kt
@@ -1,0 +1,6 @@
+package zipbap.app.global.code
+
+interface BaseCode {
+    val reasonDto: ReasonDto
+    val reasonHttpStatus: ReasonDto
+}
