+++ src/main/kotlin/zipbap/app/global/code/BaseErrorCode.kt
@@ -1,0 +1,6 @@
+package zipbap.app.global.code
+
+interface BaseErrorCode {
+    val reason: ErrorReasonDto
+    val reasonHttpStatus: ErrorReasonDto
+}
