+++ src/main/kotlin/zipbap/app/global/code/ErrorReasonDto.kt
@@ -1,0 +1,11 @@
+package zipbap.app.global.code
+
+import org.springframework.http.HttpStatus
+
+data class ErrorReasonDto(
+    val httpStatus: HttpStatus,
+    val isSuccess: Boolean,
+    val code: String,
+    val message: String
+) {
+}
