--- src/test/kotlin/zipbap/api/ZipbapApiApplicationTests.kt
@@ -1,13 +1,0 @@
-package zipbap.api
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ZipbapApiApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
