+++ src/test/kotlin/zipbap/app/ZipbapApiApplicationTests.kt
@@ -1,0 +1,13 @@
+package zipbap.app
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class ZipbapApiApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
